Implemented a cache for SkFontDescriptor.
The renderer process and gpu process manage their own cache separately. In
SkFontDescriptor::serialize(), a descriptor will be serialized as before and
stored in an unordered map with its fFamilyName as the key. Only the key will
be serialized with id=kCacheIndex if the descriptor is already cached.
In SkFontDescriptor::Deserialize, an SkFontdescriptor is deserialized as before
and then stored in an unordered map. If a kCachedIndex is received, we look for
the descriptor in the cache.

BUG: many SkFontDescriptors has empty strings as their fFamilyName. As a work
around, only descriptors with non-empty string family names are cached.
I'm looking at where the nameless font are constructed and see if there's a
way to add some identifiers there.