Implemented a simple cache for the serialization/deserialization of SkFontDescriptor: The family name and style bits are concatenated to be used as the key of the cache. Added an enum kCacheIndex = 0xFB to indicate the serialized message is a cache key. Added static SkFontDescriptor::std::unordered_map<std::string, SkFontDescriptor>   fSerializeCache to be used by SkFontDescriptor::serialize(). We first check if the font descriptor has been serialized before. If so, only the key is serialized; otherwise, the descriptor is serialized as before and then added to the cache.
Added static std::unordered_map<std::string, SkFontDescriptor> deserializeCache
in SkFontDescriptor::Deserialize(). SkFontDescriptors are deserialized as before
and then added to the cache when received. When a kCacheIndex is received, we
deserialize the key and use it to retrieve the SkFontDescriptor from the cache.

A simple comparison between the perf reports showed that this cache reduced
the time spent in SkFontDescriptor::Deserialize() by about 97%.